requirements.txt
Django==3.2
djangorestframework==3.11.0
coreapi==2.3.3
gunicorn==19.9.0
whitenoise==5.0.1
requests==2.22.0
django-environ==0.8.1
psycopg2==2.9.3
django-filter==2.2.0
celery==5.2.3
redis==4.1.0